+++ .gitignore~
@@ -1,6 +1,3 @@
-# idea why?
-.idea/
-
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
